requirements.txt
Flask==2.0.1
Flask-CORS==3.0.10
google-cloud-texttospeech==2.5.1
PyPDF2==1.26.0
pydub==0.25.1
Werkzeug==2.0.1